requirements.txt
discord.py
line-bot-sdk
